requirements.txt
flask>=2.0.0
requests>=2.28.0
google-generativeai>=0.5.0
rapidfuzz>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
